Seperated roofs, optimized walls, organized buildings.
Prepped roof models for texturing, removed the mesh collider on the stone wall prefabs, and added box colliders for the walls in city rework, city rework 1, and outside. Also organized the modular building blocks in city rework and city rework 1 so that each piece is a child of their respective macro building.